Present new service AuthorityRuleValidationService
